+++ CHANGELOG.md
@@ -1,1 +1,2 @@
 - Fix bug where `functions:secrets:set` didn't remove stale versions of a secret. (#6080)
+- Fix bug where `firebase deploy --only firestore:named-db` didn't update rules. (#6129)